requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.1
Django==4.2.7
djangorestframework==3.14.0
googletrans==4.0.0rc1
googletrans-py==4.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2023.1.1
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==2.10
pillow==10.2.0
pytz==2024.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
